+++ docs/md/releasenotes.md
@@ -19,6 +19,7 @@
 
 - Use program location as hash code for `JvmAbstractState` to allow correct use with hash sets.
 - Fix a bug in `JvmTransferRelation` handling arithmetic instructions returning category 2 values incorrectly.
+- Fix delegation of `proguard.evaluation.value.ParticularValueFactory#createReferenceValueNull()`.
 
 ### API changes